+++ sdk/spring/CHANGELOG.md
@@ -10,6 +10,12 @@
 #### Dependency Updates
 - Add dependency `com.azure.spring:spring-cloud-azure-starter-keyvault-jca`.
 
+### Spring Cloud Azure Starter Key Vault
+This section includes changes in `spring-cloud-azure-starter-keyvault` module.
+
+#### Features Added
+- Support Key Vault JCA [#44259](https://github.com/Azure/azure-sdk-for-java/pull/44259).
+
 ## 5.20.0 (2025-02-12)
 - This release is compatible with Spring Boot 3.4.0-3.4.2, 3.3.0-3.3.6, 3.2.0-3.2.12, 3.1.0-3.1.12, 3.0.0-3.0.13. (Note: 3.4.x (x>2), 3.3.y (y>6) and 3.2.z (z>12) should be supported, but they aren't tested with this release.)
 - This release is compatible with Spring Cloud 2024.0.0, 2023.0.0-2023.0.4, 2022.0.0-2022.0.5. (Note: 2024.0.x(x>0) and 2023.0.y (y>4) should be supported, but they aren't tested with this release.)